UPSTREAM: media: videobuf2-core: Rework and rename helper for request buffer count
The helper indicating whether buffers are associated with the request is
reworked and renamed to return the number of associated buffer objects.

This is useful for drivers that need to check how many buffers are in
the request to validate it.

Existing users of the helper don't need particular adaptation since the
meaning of zero/non-zero remains consistent.

BUG=chromium:902656
TEST=Checked that Cheza was working with request API enabled.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 515c5a7333be87a7d01ab267d94626a454a7e794)
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

Change-Id: I72a2fdeb6dfaadc745edd12fc0427d5ef194d3f3
Reviewed-on: https://chromium-review.googlesource.com/c/1325797
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>