UPSTREAM: udmabuf: fix error code in map_udmabuf()
We accidentally forgot to set "ret" on this error path so it means we
return NULL instead of an error pointer.  The caller checks for NULL and
changes it to an error pointer so it doesn't cause an issue at run time.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20180914065615.GA12043@mwanda
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>

(cherry picked from commit 6f19eb21a2ef9a2da938106ccc98e296ee02f4d3)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I2a55050ee659a83b9a13345e0c55508c97e8d47e
Reviewed-on: https://chromium-review.googlesource.com/1314852
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>