UPSTREAM: drm/msm/a6xx: Poll for HFI responses
The only HFI communication with the GMU on sdm845 happens
during initialization and all commands are synchronous. A fancy
interrupt tasklet and associated infrastructure is entirely
not eeded and puts us at the mercy of the scheduler.

Instead poll for the message signal and handle the response
immediately and go on our way.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>

(cherry picked from commit df0dff132905974697e2a19aa8bcc0ecc447c00e)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ib1254f8eb4c4952d436b78ea8af5ea4c3fb07a29
Reviewed-on: https://chromium-review.googlesource.com/1342181
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>