UPSTREAM: drm/amdgpu: fix the page fault of raven2
While the apg_end address is 0xffffffff, if add 1 with it, the value will be
overflow and roll back to 0. So when 0 is written to
mmMC_VM_SYSTEM_APERTURE_HIGH_ADDR, the system aperture is actually disabled. And
so any access to vram will trigger a page fault.

Raven2's HW issue only need increase the vram end address, and needn't do it on
the agp.

Signed-off-by: Huang Rui <ray.huang@amd.com>
Acked-by: Alex Deucher <alexander.deucher@amd.com>
Reviewed-by: Christian Knig <christian.koenig@amd.com>
Cc: Marek Olk <marek.olsak@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>

(cherry picked from commit 75986276217b8fcb4d5aa8fb76540fbb26c061de)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I2959fb78ea4918f6f55e2f43b385249f4b276c51
Reviewed-on: https://chromium-review.googlesource.com/1338054
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>