UPSTREAM: drm/tegra: dc: Do not register DC without primary plane
Tegra194 contains a fourth display controller that does not own any
windows. Therefore, we cannot currently assign a primary plane to it
which causes KMS to eventually crash. Do not register the display
controller if it owns no windows to work around this.

Note that we still have to enable and probe the display controller
because for some reason all display controllers need to be powered
(and/or clocked) before any registers can be accessed in any of the
display controllers.

Signed-off-by: Thierry Reding <treding@nvidia.com>

(cherry picked from commit 759d706f7cb4adbec1c222f7550dc8fa2683c089)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I8d42f04cd1b2faf9d8ddba93202ed3940a9cf586
Reviewed-on: https://chromium-review.googlesource.com/1338002
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>