UPSTREAM: drm/vmwgfx: Remove the user resource destructor check
We were checking that the resource destructor matched that of the
intended object type, to make sure the looked up resource was of the
right type.

But we already have an object type check in place which makes sure the
resource is of the right type.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Deepak Rawat <drawat@vmware.com>

(cherry picked from commit b1d05b4fc6b8f2bf5d802cb08c87d96802c4c7ba)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I933e4768bcb09e801c8b3d73b1af6f40f5299bdd
Reviewed-on: https://chromium-review.googlesource.com/1342132
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>