UPSTREAM: net: bridge: convert and rename mcast disabled
Convert mcast disabled to an option bit and while doing so convert the
logic to check if multicast is enabled instead. That is make the logic
follow the option value - if it's set then mcast is enabled and vice versa.
This avoids a few confusing places where we inverted the value that's being
set to follow the mcast_disabled logic.

Signed-off-by: Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
Reviewed-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 13cefad2f2c1d99b778c62b142e077779c5f9395)

BUG=chromium:878034
TEST=/proc/sys/net/bridge exists in crostini

Change-Id: I4ab6d73df69903c88a206f61465c2e333916957b
Signed-off-by: Stephen Barber <smbarber@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1668396
Reviewed-by: Sean Paul <seanpaul@chromium.org>