UPSTREAM: drm/vmwgfx: Don't refcount command-buffer managed resource lookups during command buffer validation
The typical pattern of these lookups are
-Lookup
-Put on validate list if not already there.
-Unreference
And since we are the exclusive user of the context during lookup time,
we can be sure that the resource will stay alive during the sequence.
So avoid taking a reference during lookup, and also avoid unreferencing
when done.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>

(cherry picked from commit 508108ea274788888408f4245438e40c90d821da)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Icfaee8acf9af4f5de6ed4cfd7d6ebfc8cda9f65a
Reviewed-on: https://chromium-review.googlesource.com/1342139
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>