UPSTREAM: drm/msm: suspend devfreq on init
Devfreq turns on and starts recommending power level as soon as it is
initialized. The GPU is still not powered on by the time the devfreq
init happens and this leads to problems on GPU's where register access
is needed to get/set power levels. So we start suspended and only restart
devfreq when GPU is powered on.

Signed-off-by: Sharat Masetty <smasetty@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>

(cherry picked from commit d3fa91c90931e6f3b26f7acbac84b44c6756fa08)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I1ce5130e60753506ef9d615d140e10a26075534d
Reviewed-on: https://chromium-review.googlesource.com/1342186
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>