UPSTREAM: MAINTAINERS: Add Maxime Ripard as drm-misc maintainer
Unfortunately Gustavo has decided to step down as drm-misc maintainer to
focus on other projects. Thanks Gustavo for your dedication and hard work!

Fortunately for us, we have a wealth of people qualified to assume a
-misc maintainer role. Maxime has done an outstanding job with sun4i and
in the community in general. I'm really excited that he agreed to take
on this responsibility and I look forward to working with him!

Cc: Dave Airlie <airlied@linux.ie>
Cc: Gustavo Padovan <gustavo@padovan.org>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Maxime Ripard <maxime.ripard@bootlin.com>
Acked-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Acked-by: Maxime Ripard <maxime.ripard@bootlin.com>
Acked-by: Dave Airlie <airlied@linux.ie>
Signed-off-by: Sean Paul <sean@poorly.run>
Link: https://patchwork.freedesktop.org/patch/msgid/20181005160256.200162-1-sean@poorly.run

(cherry picked from commit 7372fd049aa8836310f84da5f82dc9eb146915c8)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Idbda16a5b83b749725c3bfc2468a2268f0911f6b
Reviewed-on: https://chromium-review.googlesource.com/1342221
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>