UPSTREAM: Bluetooth: Remove redundant check on status
The check on status is redundant as a status has to be zero at
the point it is being checked because of a previous check and return
path via label 'unlock'.  Remove the redundant check and the deadcode
that can never be reached.

Detected by CoverityScan, CID#1471710 ("Logically dead code")

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit d17010bf6301b62df09f7878f3474f42ee0754eb)

BUG=none
TEST=none

Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Change-Id: I395c7be52087933779c3afe2a889e1d0d33a7779
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2951947
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>