UPSTREAM: mac80211: Move ieee80211_mgd_auth() EBUSY check to be before allocation
This makes it easier to conditionally replace full allocation of
auth_data to use reallocation for the case of continuing SAE
authentication. Furthermore, there was not really any point in having
this check done so late in the function after having already completed
number of steps that cannot be used anyway in the error case.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 8d7432a2f53bc283148062955cbe7ffbf713e646)

BUG=b:172212319
TEST=suite:wifi_matfunc, wpa3 tests
TEST=`tast run -var network.HostapHwsim.runArgs='-f sae' localhost:9222
     network.HostapHwsim.full`

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I6c154749165375f3719fe8a14283d80c69fd5380
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2718949
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>