CHROMIUM: config: Build in i2c_chardev on qualcomm
We want to do this for reasons similar to those expressed in CL:366000
("CHROMIUM: config: Build in i2c_hid and i2c_chardev on rockchip").
Basically: i2c_chardev is critical path for bootup and needs to be all
taken care of before UI starts up.

This is the equivalent of:
  echo 'CONFIG_I2C_CHARDEV=y' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

arm64-generic config was not updated in this case since we'd prefer a
smaller kernel size vs. quicker bootup on that board.

BUG=b:117576582
TEST=modprobe configs; zgrep "I2C_CHARDEV" /proc/config.gz

Change-Id: Id6c6a0192baf480ef9b1764da2911bb8b5710177
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1277370
Reviewed-by: Guenter Roeck <groeck@chromium.org>