UPSTREAM: drm: panel-orientation-quirks: Add quirk for Acer One 10 (S1003)
The Acer One 10 uses a clamshell design with a detachable keyboard.
As such in normal operating mode, with the keyboard attach the device
is in landscape mode (and the Acer logo at boot also shows in landscape
mode).

But the device uses a portrait screen rotated 90 degrees (sigh). This
commit adds a quirk for this device so that we shown the fbcon the
right way up and that we hint userspace to also show e.g. plymouth and
gdm the right way up.

Cc: stable@vger.kernel.org
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20181012101610.29100-1-hdegoede@redhat.com

(cherry picked from commit 0e8afefd5da4875ddea9aa4ad17a2540a2bf9736)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I1b4d829627c44b5f7b406f0c43de536cf1cd9a41
Reviewed-on: https://chromium-review.googlesource.com/1342307
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>