CHROMIUM: iwl7000: remove FTM
Clean up FTM/TOF, in the sense of restoring it to the state
that's currently upstream. After merging back the new APIs
from upstream, we'll restore the code to work over those.

Note that there are some minor differences to upstream to
avoid errors in kernel-doc, we used to not check that.

Change-Id: I0d872dba5c41e769c4a5c9bcd3d78a195609b012
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 1ec7aa43cf884548840fd8ce8f0bbfc7b8570b3b