UPSTREAM: KVM: VMX: enable nested virtualization by default
With live migration support and finally a good solution for exception
event injection, nested VMX should be ready for having a stable userspace
ABI.  The results of syzkaller fuzzing are not perfect but not horrible
either (and might be partially due to running on GCE, so that effectively
we're testing three-level nesting on a fork of upstream KVM!).  Enabling
it by default seems like a nice way to conclude the 4.20 pull request. :)

Unfortunately, enabling nested SVM in 2009 (commit 4b6e4dca701) was a
bit premature.  However, until live migration support is in place we can
reasonably expect that it does not offer much in terms of ABI guarantees.
Therefore we are still in time to break things and conform as much as
possible to the interface used for VMX.

BUG=chromium:993253
TEST=Check that /dev/kvm appears in the guest.

Suggested-by: Jim Mattson <jmattson@google.com>
Suggested-by: Liran Alon <liran.alon@oracle.com>
Reviewed-by: Liran Alon <liran.alon@oracle.com>
Celebrated-by: Liran Alon <liran.alon@oracle.com>
Celebrated-by: Wanpeng Li <kernellwp@gmail.com>
Celebrated-by: Wincy Van <fanwenyi0529@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 1e58e5e59148916fa43444a406335a990783fb78)
Signed-off-by: Abhishek Bhardwaj <abhishekbh@google.com>

Change-Id: I6c5fdec1c8baf3cc9f75aa94594109a4cbe74e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1929075
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>