FROMGIT: drm/msm/hdmi: Enable HPD after HDMI IRQ is set up
SoCs that contain MDP5 have a top level wrapper called MDSS that
manages locks, power and irq for the sub-blocks within it.

Irq for HDMI is also routed through the MDSS.

Shortly after the Hot Plug Detection (HPD) is enabled in HDMI,
HDMI interrupts are recieved by the MDSS interrupt handler.
However at this moment the HDMI irq is still not mapped to
the MDSS irq domain so the HDMI irq handler cannot be called
to process the interrupts.

This leads to a flood of HDMI interrupts on CPU 0.

If we are lucky to have the HDMI initialization running on a
different CPU, it will eventually map the HDMI irq to MDSS irq
domain, the next HDMI interrupt will be handled by the HDMI irq
handler, the interrupt flood will stop and we will recover.

If the HDMI initialization is running on CPU 0, then it cannot
complete and there is nothing to stop the interrupt flood on
CPU 0. The system is stuck.

Fix this by moving the HPD enablement after the HDMI irq is
mapped to the MDSS irq domain.

Signed-off-by: Todor Tomov <todor.tomov@linaro.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>

(cherry picked from commit 66b1e71547aa9b681ac5274598e36490e9be21b0)
(am from git://people.freedesktop.org/~robclark/linux)

BUG=None
TEST=Tested on cheza

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ib622934794c9019282fdb5dfa253ba1f8c65b6b2
Reviewed-on: https://chromium-review.googlesource.com/1350003
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>