CHROMIUM: iwl7000: mvm: implement CSI reporting
Implement CSI (channel estimation matrix) reporting in the mvm
driver, if the firmware has the capability.

Currently only a debugfs API is provided as the API is still
under discussion.

For now, RX aggregation must be disabled to use this feature
on data frames as we haven't found a good way to attach the
data to A-MPDUs, given complexities with multi-queue.

Change-Id: Ib49b72f389ed95537db10f6e7cab271060fd3b94
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 45b300f43f845c477fdd201517ab836d4c72c60e