FROMGIT: dt-bindings: drm/panel: simple: Add no-hpd property
Some eDP panels that are designed to be always connected to a board
use their HPD signal to signal that they've finished powering on and
they're ready to be talked to.

However, for various reasons it's possible that the HPD signal from
the panel isn't actually hooked up.  In the case where the HPD isn't
hooked up you can look at the timing diagram on the panel datasheet
and insert a delay for the maximum amount of time that the HPD might
take to come up.

Let's add a property in the device tree for this concept.

Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20181025222134.174583-1-dianders@chromium.org
(cherry picked from commit cb028e49129f352d2f2645727ceeeacdd64761f4
 https://github.com/freedesktop/drm-misc.git for-linux-next)

BUG=b:112356283
TEST=Display still works

Change-Id: I3b16fb68ea5f34fbad5c09959876e8984e0e94e5
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1297444
Reviewed-by: Sean Paul <seanpaul@chromium.org>