FROMGIT: drm/panel: simple: Add "no-hpd" delay for Innolux TV123WAM
If the HPD signal isn't hooked up to this panel we need a 200 ms
delay.  In the datasheet this is shown as the maximum time that HPD
will take to be asserted after power is given to the panel.

Reviewed-by: Sean Paul <sean@poorly.run>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20181025222134.174583-3-dianders@chromium.org
(cherry picked from commit 625d3b5c2ea612e4adf8b289384ad654c207c2bb
 https://github.com/freedesktop/drm-misc.git for-linux-next)

BUG=b:112356283
TEST=Display still works

Change-Id: I44053613eeaa2aad3c831bef6f50c6b78b255bef
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1297446
Reviewed-by: Sean Paul <seanpaul@chromium.org>