FROMGIT: OPP: Add dev_pm_opp_xlate_performance_state() helper
dev_pm_genpd_set_performance_state() needs to handle performance state
propagation going forward. Currently this routine only gets the required
performance state of the device's genpd as an argument, but it doesn't
know how to translate that to master genpd(s) of the device's genpd.

Introduce a new helper dev_pm_opp_xlate_performance_state() which will
be used to translate from performance state of a device (or genpd
sub-domain) to another device (or master genpd).

Normally the src_table (of genpd sub-domain) will have the
"required_opps" property set to point to one of the OPPs in the
dst_table (of master genpd), but in some cases the genpd and its master
have one to one mapping of performance states and so none of them have
the "required-opps" property set. Return the performance state of the
src_table as it is in such cases.

Tested-by: Rajendra Nayak <rnayak@codeaurora.org>
Reviewed-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit c8a59103e22b191e363fc0a90e08515a915b278d
 git://git.kernel.org/pub/scm/linux/kernel/git/vireshk/pm.git opp/linux-next)

BUG=b:80592702
TEST=Build and boot

Change-Id: I3fdb576257ecb0835f3006ba4710e5a03ceddc8a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1372622
Reviewed-by: Stephen Boyd <swboyd@chromium.org>